Object avoidance states implementation and code clean up
Implemented 2 states of object avoidance (reorientation of the car and bug algorithm). Clean up some old codes with object avoidance substates, related variable.

Object avoidance now has PD control on the motor speed. The speed is dependent on the measurement of the ultrasonic sensor and change in measurement. Would be better if the derivative term consider the changes in yaw angle measurement from the IMU